+++ packages/checklist/lib/checklist.dart
@@ -1,11 +1,12 @@
-// @MappableLib(
-//   generateInitializerForScope: InitializerScope.package,
-// )
+@MappableLib(
+  generateInitializerForScope: InitializerScope.package,
+)
 library checklist;
 
 import 'package:dart_mappable/dart_mappable.dart';
 
-// export 'checklist.init.dart';
+//exporting all the files from checklist package
+export 'checklist.init.dart';
 export 'models/entities/service.dart';
 export 'models/entities/service_attributes.dart';
 export 'models/entities/service_definition.dart';
@@ -15,7 +16,9 @@ export 'data/repositories/local/service_definition.dart';
 export 'data/repositories/remote/service.dart';
 export 'data/repositories/remote/service_definition.dart';
 export 'data/repositories/remote/service_attributes.dart';
-export 'models/entities/attributes.dart';
 export 'utils/utils.dart';
+export 'blocs/service.dart';
+export 'blocs/service_definition.dart';
+export 'blocs/app_localization.dart';