+++ packages/checklist/lib/models/entities/service.dart
@@ -3,9 +3,7 @@ import 'package:dart_mappable/dart_mappable.dart';
 import 'package:drift/drift.dart';
 
 import 'package:digit_data_model/data_model.dart';
-
-// import '../../data/local_store/sql_store/sql_store.dart';
-import '../entities/service_attributes.dart' as svc_att;
+import '../entities/service_attributes.dart';
 
 part 'service.mapper.dart';
 
@@ -56,7 +54,7 @@ class ServiceModel extends EntityModel with ServiceModelMappable {
   final bool? nonRecoverableError;
   final String? tenantId;
   final int? rowVersion;
-  final List<svc_att.ServiceAttributesModel>? attributes;
+  final List<ServiceAttributesModel>? attributes;
   final ServiceAdditionalFields? additionalFields;
 
   ServiceModel({
@@ -72,10 +70,13 @@ class ServiceModel extends EntityModel with ServiceModelMappable {
     this.tenantId,
     this.rowVersion,
     this.attributes,
-    super.auditDetails,super.clientAuditDetails,
+    super.auditDetails,
+    super.clientAuditDetails,
     super.isDeleted = false,
   }): super();
 
+
+  //Helper object to represents the data you want to insert or update in a table
   ServiceCompanion get companion {
     return ServiceCompanion(
       auditCreatedBy: Value(auditDetails?.createdBy),