+++ packages/checklist/lib/utils/extensions/context_utility.dart
@@ -1,10 +1,14 @@
-part of 'extensions.dart';
+import 'package:digit_data_model/data_model.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import '../utils.dart';
 
 extension ContextUtilityExtensions on BuildContext {
   int millisecondsSinceEpoch([DateTime? dateTime]) {
     return (dateTime ?? DateTime.now()).millisecondsSinceEpoch;
   }
 
+  // Provides Checklist repository based on the persistence configuration
   DataRepository<D, R>
   repository<D extends EntityModel, R extends EntitySearchModel>(
       BuildContext context,