+++ packages/digit_data_model/CHANGELOG.md
@@ -1,12 +1,21 @@
 ## 1.0.4-dev.2
 
-*Removed Service and Service Definition models
-   - Removed Service repositories and Service blocs to integrate in checklist package
+* Removed Service and Service Definition models
+  - Removed Service repositories and Service blocs to integrate in checklist package
 
+## 1.0.3
+
+* Added localization table
+* Updated individual local search results with user uuid mapping
+
+## 1.0.3-dev.4
+
+* Updated individual local search results with user uuid mapping
+  
 ## 1.0.3-dev.3
 
-* Created localization table
-    - Creation of the localization table.
+* Created localization table 
+  - Creation of the localization table.
 
 ## 1.0.3-dev.2